+++ app/javascript/src/admin/micro-clusters/Spinner.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const Spinner = () => (
+  <div className="loader">
+    <i className="fa fa-spin fa-refresh" />
+  </div>
+);
